feat: verified store repository 분리
