+++ KnockKnock-iOS/Repository/HotPostRepository.swift
@@ -19,26 +19,6 @@ protocol HotPostRepositoryProtocol {
 
 final class HotPostRepository: HotPostRepositoryProtocol {
 
-  /// 인증 스토어 목록 조회
-  func requestVerifiedStoreList() async -> [Store]? {
-    do {
-      let result = try await KKNetworkManager
-        .shared
-        .asyncRequest(
-          object: ApiResponseDTO<[StoreDTO]>.self,
-          router: .getHomeVerificationShop
-        )
-
-      guard let data = result.data else { return nil }
-      return data.map { $0.toDomain() }
-
-    } catch let error {
-      print(error)
-
-      return nil
-    }
-  }
-
   /// 인기 게시글 조회
   func requestHotPost(
     challengeId: Int,
