+++ KnockKnock-iOS/Scenes/Home/HomeWorker.swift
@@ -12,20 +12,25 @@ protocol HomeWorkerProtocol {
     challengeId: Int,
     completionHandler: @escaping ([HotPost]) -> Void
   )
-  func fetchChallengeList(completionHandler: @escaping([ChallengeTitle]) -> Void)
+  func fetchChallengeList(
+    completionHandler: @escaping([ChallengeTitle]) -> Void
+  )
   func fetchEventList() async -> [Event]
   func fetchVerifiedStore() async -> [Store]?
 }
 
 final class HomeWorker: HomeWorkerProtocol {
 
+  private let verifiedStoreRepository: VerifiedStoreRepositoryProtocol
   private let hotPostRepository: HotPostRepositoryProtocol
   private let eventRepository: EventRepositoryProtocol
 
   init(
+    verifiedStoreRepository: VerifiedStoreRepositoryProtocol,
     hotPostRepository: HotPostRepositoryProtocol,
     eventRepository: EventRepositoryProtocol
   ) {
+    self.verifiedStoreRepository = verifiedStoreRepository
     self.hotPostRepository = hotPostRepository
     self.eventRepository = eventRepository
   }
@@ -53,7 +58,7 @@ final class HomeWorker: HomeWorkerProtocol {
   }
 
   func fetchVerifiedStore() async -> [Store]? {
-    return await self.homeRepository.requestVerifiedStoreList()
+    return await self.verifiedStoreRepository.requestVerifiedStoreList()
   }
 
   func fetchEventList() async -> [Event] {
